[refactor] Move `NetworkImage` to single file
