+++ Features/Sources/Features/ThemeBoxCardView.swift
@@ -41,13 +41,3 @@ public struct ThemeBoxCardView: View {
     }
   }
 }
-
-struct NetworkImage: View {
-  let url: URL?
-  var body: some View {
-    KFImage(url)
-      .resizable()
-      .aspectRatio(contentMode: .fill)
-      .frame(maxWidth: .infinity, maxHeight: .infinity)
-  }
-}
